Create CONTRIBUTING.md
I have added a few initial thoughts but by no means is this comprehensive. The first section is a general statement of inclusion and the second section is a more solid guideline on how to submit. The submission guidelines are purely based on my own thoughts and at this stage only meant to spark discussion, not be followed rigourously.